# Perf improvements to char.IsWhiteSpace / char.IsUpper / char.IsLower