@EgorBo 2019-09-23
This PR also was supposed to remove `CategoryForLatin1` bound checks: https://github.com/dotnet/coreclr/pull/25912