@GrabYourPitchforks 2019-09-24
Unit tests at https://github.com/dotnet/corefx/pull/41265.